% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeResults.R
\name{analyzeNormalizations}
\alias{analyzeNormalizations}
\title{Calculate measures for normalization results}
\usage{
analyzeNormalizations(nr, comparisons = NULL, categoricalAnova = FALSE,
  varFilterFrac = NULL)
}
\arguments{
\item{nr}{Normalyzer results object with calculated results.}

\item{comparisons}{Target sample contrasts to run.}

\item{categoricalAnova}{ANOVA can be categorical or numerical.}

\item{varFilterFrac}{Perform variance filtering before tests.}
}
\value{
Normalyzer results with attached evaluation results object.
}
\description{
Calculate measures for normalization results
}
\examples{
data(example_data)
data(example_design)
normObj <- getVerifiedNormalyzerObject("job_name", example_design, example_data)
normResults <- normMethods(normObj)
normResultsWithEval <- analyzeNormalizations(normResults)
}
