% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerResults.R
\docType{methods}
\name{basicMetricNormalizations}
\alias{basicMetricNormalizations}
\alias{basicMetricNormalizations,NormalyzerResults-method}
\title{Generate basic metrics normalizations}
\usage{
basicMetricNormalizations(nr)

\S4method{basicMetricNormalizations}{NormalyzerResults}(nr)
}
\arguments{
\item{nr}{Normalyzer results object.}
}
\value{
None
}
\description{
Generate basic metrics normalizations
}
\keyword{internal}
