% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeResults.R
\name{calculateAvgMadMem}
\alias{calculateAvgMadMem}
\title{Calculate average MAD (Median Absolute Deviation) for each feature in
each condition and then calculates the average for each replicate group}
\usage{
calculateAvgMadMem(methodList, sampleReplicateGroups)
}
\arguments{
\item{methodList}{List containing normalized matrices.}

\item{sampleReplicateGroups}{Condition header.}
}
\value{
condAvgMadMat Matrix with average MAD for each biological condition.
}
\description{
Calculate average MAD (Median Absolute Deviation) for each feature in
each condition and then calculates the average for each replicate group
}
\keyword{internal}
