% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeResults.R
\name{calculateAvgReplicateVariation}
\alias{calculateAvgReplicateVariation}
\title{Calculate average variance for each feature in each condition and then
calculate the average for each replicate group}
\usage{
calculateAvgReplicateVariation(methodList, sampleReplicateGroups)
}
\arguments{
\item{methodList}{List containing normalized matrices.}

\item{sampleReplicateGroups}{Condition header.}
}
\value{
avgVarianceMat Matrix with average variance for each biological
condition
}
\description{
Calculate average variance for each feature in each condition and then
calculate the average for each replicate group
}
\keyword{internal}
