% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerEvaluationResults.R
\docType{methods}
\name{calculateAvgVar}
\alias{calculateAvgVar}
\alias{calculateAvgVar,NormalyzerEvaluationResults-method}
\title{Calculate average variation measures}
\usage{
calculateAvgVar(ner, nr)

\S4method{calculateAvgVar}{NormalyzerEvaluationResults}(ner, nr)
}
\arguments{
\item{ner}{NormalyzerDE evaluation object.}

\item{nr}{NormalyzerDE results object.}
}
\value{
None
}
\description{
Calculate average variation measures
}
\keyword{internal}
