% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeResults.R
\name{calculateCorrelations}
\alias{calculateCorrelations}
\title{Pearson and Spearman correlation calculations for methods and samples
Calculates internal correlation per condition}
\usage{
calculateCorrelations(nr, ner)
}
\arguments{
\item{nr}{Normalyzer results object with calculated results.}

\item{ner}{Normalyzer evaluation object.}
}
\value{
Normalyzer evaluation object with attached evaluation results.
}
\description{
Pearson and Spearman correlation calculations for methods and samples
Calculates internal correlation per condition
}
\keyword{internal}
