% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeResults.R
\name{calculateCorrSum}
\alias{calculateCorrSum}
\title{Calculates internal correlations for each condition having at least two
samples and returns a vector with correlation values corresponding to each
condition}
\usage{
calculateCorrSum(methodData, allReplicateGroups, sampleGroupsWithReplicates,
  corrType)
}
\arguments{
\item{methodData}{Expression data matrix}

\item{allReplicateGroups}{Full condition header corresponding to data tables
columns}

\item{sampleGroupsWithReplicates}{Unique conditions where number of
replicates exceeds one}

\item{corrType}{Type of correlation (Pearson or Spearman)}
}
\value{
corSums
}
\description{
Calculates internal correlations for each condition having at least two
samples and returns a vector with correlation values corresponding to each
condition
}
\examples{
data("example_data_only_values")
data("example_design")
testDesign <- example_design[which(example_design$group \%in\% c("1", "2", "3")), ]
testData <- example_data_only_values[, as.character(testDesign$sample)]
groupHeader <- testDesign$group
uniqueGroups <- unique(groupHeader)
calculateCorrSum(testData, groupHeader, uniqueGroups, "pearson")
}
