% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerEvaluationResults.R
\docType{methods}
\name{calculateCV}
\alias{calculateCV}
\alias{calculateCV,NormalyzerEvaluationResults-method}
\title{Calculate coefficient of variation (relative standard deviation)
Stores percentage variation compared to}
\usage{
calculateCV(ner, nr)

\S4method{calculateCV}{NormalyzerEvaluationResults}(ner, nr)
}
\arguments{
\item{ner}{NormalyzerDE evaluation object.}

\item{nr}{NormalyzerDE results object.}
}
\value{
None
}
\description{
Calculate coefficient of variation (relative standard deviation)
Stores percentage variation compared to
}
\keyword{internal}
