% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeResults.R
\name{calculateFeatureCV}
\alias{calculateFeatureCV}
\title{Calculate CV values for each feature. Iterates through each normalization 
method and calculates a matrix of CV values where each column correspond to 
a method and each row corresponds to a feature.}
\usage{
calculateFeatureCV(methodList)
}
\arguments{
\item{methodList}{List containing normalized matrices.}
}
\value{
methodFeatureCVMatrix Matrix with feature as rows and normalization
  method as columns
}
\description{
Calculate CV values for each feature. Iterates through each normalization 
method and calculates a matrix of CV values where each column correspond to 
a method and each row corresponds to a feature.
}
\keyword{internal}
