% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerEvaluationResults.R
\docType{methods}
\name{calculateMAD}
\alias{calculateMAD}
\alias{calculateMAD,NormalyzerEvaluationResults-method}
\title{Calculate median absolute deviation measures}
\usage{
calculateMAD(ner, nr)

\S4method{calculateMAD}{NormalyzerEvaluationResults}(ner, nr)
}
\arguments{
\item{ner}{NormalyzerDE evaluation object.}

\item{nr}{Normalyzer results object.}
}
\value{
None
}
\description{
Calculate median absolute deviation measures
}
\keyword{internal}
