% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerStatistics.R
\docType{methods}
\name{calculatePairwiseComparisonsLimma}
\alias{calculatePairwiseComparisonsLimma}
\alias{calculatePairwiseComparisonsLimma,NormalyzerStatistics,character,character-method}
\title{Calculate Limma comparisons between target samples}
\usage{
calculatePairwiseComparisonsLimma(nst, comparisons, condCol, splitter = "-")


  \S4method{calculatePairwiseComparisonsLimma}{NormalyzerStatistics,character,character}(nst,
  comparisons, condCol, splitter = "-")
}
\arguments{
\item{nst}{NormalyzerDE statistics object}

\item{comparisons}{Vector containing statistical contrasts.}

\item{condCol}{Column name in design matrix containing condition values}

\item{splitter}{String splitting the comparisons into condition levels}
}
\value{
None
}
\description{
Calculate Limma comparisons between target samples
}
\keyword{internal}
