% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeResults.R
\name{calculatePercentageAvgDiffInMat}
\alias{calculatePercentageAvgDiffInMat}
\title{General function for calculating percentage difference of average column
means in matrix}
\usage{
calculatePercentageAvgDiffInMat(targetMat)
}
\arguments{
\item{targetMat}{Matrix for which column means should be compared}
}
\value{
percDiffVector Vector with percentage difference, where first element
  always will be 100
}
\description{
General function for calculating percentage difference of average column
means in matrix
}
\keyword{internal}
