% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeResults.R
\name{calculateReplicateCV}
\alias{calculateReplicateCV}
\title{Calculate CV per replicate group and normalization technique}
\usage{
calculateReplicateCV(methodList, sampleReplicateGroups)
}
\arguments{
\item{methodList}{List containing normalized matrices.}

\item{sampleReplicateGroups}{Condition header.}
}
\value{
avgCVPerNormAndReplicates Matrix with group CVs as rows and
  normalization technique as columns
}
\description{
Calculate CV per replicate group and normalization technique
}
\keyword{internal}
