% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerEvaluationResults.R
\docType{methods}
\name{calculateSignificanceMeasures}
\alias{calculateSignificanceMeasures}
\alias{calculateSignificanceMeasures,NormalyzerEvaluationResults-method}
\title{Calculate significance measures (p-value and FDR-value) for ANOVA and
Kruskal-Wallis}
\usage{
calculateSignificanceMeasures(ner, nr, categoricalAnova, varFilterFrac)

\S4method{calculateSignificanceMeasures}{NormalyzerEvaluationResults}(ner, nr,
  categoricalAnova = FALSE, varFilterFrac = NULL)
}
\arguments{
\item{ner}{NormalyzerDE evaluation object.}

\item{nr}{Normalyzer results object.}

\item{categoricalAnova}{Categorical groupwise comparison instead of numeric}

\item{varFilterFrac}{Variance filter high variance data}
}
\value{
None
}
\description{
Calculate significance measures (p-value and FDR-value) for ANOVA and
Kruskal-Wallis
}
\keyword{internal}
