% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createDirectory}
\alias{createDirectory}
\title{Create directory, or return error if already present}
\usage{
createDirectory(targetPath)
}
\arguments{
\item{targetPath}{Path where to attempt to create directory}
}
\value{
None
}
\description{
Create directory, or return error if already present
}
\keyword{internal}
