% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{defaultInputOptions}
\alias{defaultInputOptions}
\title{Create input options for default (delimited) matrices}
\usage{
defaultInputOptions(sep = "\\t")
}
\arguments{
\item{sep}{Field separator used when reading the data matrix.}
}
\value{
A list suitable to pass as \code{inputOptions}.
}
\description{
Helper to construct an \code{inputOptions} list for \code{inputFormat =
"default"}. Currently this supports selecting the input delimiter.
}
\examples{
defaultInputOptions(sep = ",")
}
