% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerDataset.R
\docType{methods}
\name{detectSingleReplicate}
\alias{detectSingleReplicate}
\alias{detectSingleReplicate,NormalyzerDataset-method}
\title{Detect single replicate, and assign related logical}
\usage{
detectSingleReplicate(nds)

\S4method{detectSingleReplicate}{NormalyzerDataset}(nds)
}
\arguments{
\item{nds}{Normalyzer dataset}
}
\value{
bool on whether sample contains only one sample group
}
\description{
Detect single replicate, and assign related logical
}
\keyword{internal}
