% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{elapsedSecondsBetweenSystimes}
\alias{elapsedSecondsBetweenSystimes}
\title{Get number of seconds between two Sys.time() objects}
\usage{
elapsedSecondsBetweenSystimes(start, end)
}
\arguments{
\item{start}{Start-time object}

\item{end}{End-time object}
}
\value{
None
}
\description{
Get number of seconds between two Sys.time() objects
}
\keyword{internal}
