% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Small example dataset}
\format{A data frame containing annotation and expression data}
\usage{
example_data
}
\description{
Small example dataset
}
\keyword{datasets}
