% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data_only_values}
\alias{example_data_only_values}
\title{Same data as in "example_data", but omitting the annotation meaning that it
only contains the expression data.}
\format{
A data frame containing expression data
}
\usage{
example_data_only_values
}
\description{
Same data as in "example_data", but omitting the annotation meaning that it
only contains the expression data.
}
\keyword{internal}
