% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_design}
\alias{example_design}
\title{Small example design matrix}
\format{A data frame containing expression data}
\usage{
example_design
}
\description{
Small example design matrix
}
\keyword{datasets}
