% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_stat_data}
\alias{example_stat_data}
\title{Same data as in "example_data", but normalized and ready for statistical
processing.}
\format{A normalized data frame ready for statistical processing}
\usage{
example_stat_data
}
\description{
Same data as in "example_data", but normalized and ready for statistical
processing.
}
\keyword{internal}
