% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_stat_summarized_experiment}
\alias{example_stat_summarized_experiment}
\title{SummarizedExperiment object prepared with design-matrix, data-matrix
and annotation columns for normalized data}
\format{
An instance of the class SummarizedExperiment with stats data
}
\usage{
example_stat_summarized_experiment
}
\description{
SummarizedExperiment object prepared with design-matrix, data-matrix
and annotation columns for normalized data
}
\keyword{internal}
