% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_summarized_experiment}
\alias{example_summarized_experiment}
\title{SummarizedExperiment object prepared with design-matrix, data-matrix
and annotation columns loaded}
\format{An instance of the class SummarizedExperiment}
\usage{
example_summarized_experiment
}
\description{
SummarizedExperiment object prepared with design-matrix, data-matrix
and annotation columns loaded
}
\keyword{internal}
