% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_wide_data}
\alias{example_wide_data}
\title{Full raw NormalyzerDE matrix used for internal testing}
\format{A data table ready for analysis in NormalyzerDE}
\usage{
example_wide_data
}
\description{
Full raw NormalyzerDE matrix used for internal testing
}
\keyword{internal}
