% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_wide_design}
\alias{example_wide_design}
\title{Design matrix belonging together with example_wide_data. Used for
internal testing.}
\format{
A design table ready for analysis in NormalyzerDE
}
\usage{
example_wide_design
}
\description{
Design matrix belonging together with example_wide_data. Used for
internal testing.
}
\keyword{internal}
