% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeResults.R
\name{findLowlyVariableFeaturesCVs}
\alias{findLowlyVariableFeaturesCVs}
\title{Uses a list of FDR-values to extract features with low variance in the
log2-transformed dataset. This is then used to calculate the average CV
for these 'lowly variable' features in each normalization approach}
\usage{
findLowlyVariableFeaturesCVs(referenceFDR, methodList)
}
\arguments{
\item{referenceFDR}{List of FDR values used as non-normalized reference}

\item{methodList}{List containing normalized matrices}
}
\value{
lowVarFeaturesAverageCVs Average CV values for lowly variable
features in each normalization approach
}
\description{
Uses a list of FDR-values to extract features with low variance in the
log2-transformed dataset. This is then used to calculate the average CV
for these 'lowly variable' features in each normalization approach
}
\keyword{internal}
