% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateStatistics.R
\name{generateAnnotatedMatrix}
\alias{generateAnnotatedMatrix}
\title{Generate annotated dataframe from statistics object}
\usage{
generateAnnotatedMatrix(nst)
}
\arguments{
\item{nst}{NormalyzerDE statistics object.}
}
\value{
outDf Annotated statistics matrix
}
\description{
Generate annotated dataframe from statistics object
}
\examples{
data(example_stat_data)
data(example_design)
statObj <- setupStatisticsObject(example_design, example_stat_data)
statObj <- calculateContrasts(statObj, comparisons=c("1-2", "2-3"), condCol="group", type="limma")
annotDf <- generateAnnotatedMatrix(statObj)
}
