% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{generateNormalyzerDataset}
\alias{generateNormalyzerDataset}
\title{Setup Normalyzer dataset from given raw data}
\usage{
generateNormalyzerDataset(jobName, designMatrix, fullRawMatrix, sampleNameCol,
  groupNameCol, quiet = FALSE)
}
\arguments{
\item{jobName}{Name of ongoing run.}

\item{designMatrix}{Dataframe containing condition matrix.}

\item{fullRawMatrix}{Dataframe with unparsed input data.}

\item{sampleNameCol}{Name of column in design matrix containing sample names.}

\item{groupNameCol}{Name of column in design matrix contaning conditions.}
}
\value{
Data object representing loaded data.
}
\description{
Setup Normalyzer dataset from given raw data
}
\keyword{internal}
