% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normMethods.R
\name{generateNormalyzerResultsObject}
\alias{generateNormalyzerResultsObject}
\title{Create empty Normalyzer results object from Normalyzer data object}
\usage{
generateNormalyzerResultsObject(nds)
}
\arguments{
\item{nds}{Normalyzer dataset object.}
}
\value{
Empty Normalyzer results object.
}
\description{
Create empty Normalyzer results object from Normalyzer data object
}
\examples{
data(example_data)
data(example_design)
normObj <- getVerifiedNormalyzerObject("job_name", example_design, example_data)
normResObj <- generateNormalyzerResultsObject(normObj)
}
\keyword{internal}
