% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higherOrderNormMethods.R
\name{getCombinedMatrix}
\alias{getCombinedMatrix}
\title{Merge multiple dataframes using provided function}
\usage{
getCombinedMatrix(mList, combFunc)
}
\arguments{
\item{mList}{List containing dataframes of same shape}

\item{combFunc}{Function performing elementwise merge of matrices}
}
\value{
combinedMatrix A single dataframe with combined data
}
\description{
Merge multiple dataframes using provided function
}
\keyword{internal}
