% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getLowNALinesContrast}
\alias{getLowNALinesContrast}
\title{Get contrast vector (TRUE/FALSE-values) indicating whether at least half
of values in row have non-NA values}
\usage{
getLowNALinesContrast(dataMatrix, replicateHeader)
}
\arguments{
\item{dataMatrix}{Matrix with expression values for entities in replicate 
samples.}

\item{replicateHeader}{Header showing how samples in matrix are replicated.}
}
\value{
Contrast vector
}
\description{
Get contrast vector (TRUE/FALSE-values) indicating whether at least half
of values in row have non-NA values
}
\keyword{internal}
