% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getLowRepCountFilterContrast}
\alias{getLowRepCountFilterContrast}
\title{Get contrast vector (TRUE/FALSE-values) indicating whether all samples
specified in replicateHeader have at least one non-NA value}
\usage{
getLowRepCountFilterContrast(dataMatrix, replicateHeader, minCount = 1)
}
\arguments{
\item{dataMatrix}{Matrix with expression values for entities in replicate 
samples.}

\item{replicateHeader}{Header showing how samples in matrix are replicated.}

\item{minCount}{Minimum number of non-NA required for sample.}
}
\value{
Contrast vector
}
\description{
Get contrast vector (TRUE/FALSE-values) indicating whether all samples
specified in replicateHeader have at least one non-NA value
}
\keyword{internal}
