% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerResults.R
\docType{methods}
\name{getNormalizationMatrices}
\alias{getNormalizationMatrices}
\alias{getNormalizationMatrices,NormalyzerResults-method}
\title{Get list with normalization matrices}
\usage{
getNormalizationMatrices(nr)

\S4method{getNormalizationMatrices}{NormalyzerResults}(nr)
}
\arguments{
\item{nr}{Normalyzer results object.}
}
\value{
None
}
\description{
Get list with normalization matrices
}
\keyword{internal}
