% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getReplicateSortedData}
\alias{getReplicateSortedData}
\title{Get dataframe with raw data column sorted on replicates}
\usage{
getReplicateSortedData(rawDataOnly, groups)
}
\arguments{
\item{rawDataOnly}{Data frame with unparsed input data matrix.}

\item{groups}{Vector containing condition levels.}
}
\value{
rawData sorted on replicate
}
\description{
Get dataframe with raw data column sorted on replicates
}
\keyword{internal}
