% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerResults.R
\docType{methods}
\name{getSlotNameList}
\alias{getSlotNameList}
\alias{getSlotNameList,NormalyzerResults-method}
\title{Get list of names for slots}
\usage{
getSlotNameList(nr)

\S4method{getSlotNameList}{NormalyzerResults}(nr)
}
\arguments{
\item{nr}{Normalyzer results object.}
}
\value{
None
}
\description{
Get list of names for slots
}
\keyword{internal}
