% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerResults.R
\docType{methods}
\name{getUsedMethodNames}
\alias{getUsedMethodNames}
\alias{getUsedMethodNames,NormalyzerResults-method}
\title{Get vector of labels for used methods}
\usage{
getUsedMethodNames(nr)

\S4method{getUsedMethodNames}{NormalyzerResults}(nr)
}
\arguments{
\item{nr}{Normalyzer results object.}
}
\value{
None
}
\description{
Get vector of labels for used methods
}
\keyword{internal}
