% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normMethods.R
\name{globalIntensityNormalization}
\alias{globalIntensityNormalization}
\title{Normalization adjusting based on total sample intensity}
\usage{
globalIntensityNormalization(rawMatrix)
}
\arguments{
\item{rawMatrix}{Target matrix to be normalized}
}
\value{
Normalized and log-transformed matrix
}
\description{
Normalization adjusting based on total sample intensity
}
\examples{
data(example_data_only_values)
normMatrix <- globalIntensityNormalization(example_data_only_values)
}
