% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerResults.R
\docType{methods}
\name{initializeResultsObject}
\alias{initializeResultsObject}
\alias{initializeResultsObject,NormalyzerResults-method}
\title{Initialize Normalyzer results object}
\usage{
initializeResultsObject(nr)

\S4method{initializeResultsObject}{NormalyzerResults}(nr)
}
\arguments{
\item{nr}{Normalyzer results object.}
}
\value{
None
}
\description{
Initialize Normalyzer results object
}
\keyword{internal}
