% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{loadData}
\alias{loadData}
\title{Load raw data into dataframe}
\usage{
loadData(dataPath, inputFormat = "default")
}
\arguments{
\item{dataPath}{File path to data matrix.}

\item{inputFormat}{If input is given in standard NormalyzerDE format, 
Proteios format or in MaxQuant protein or peptide format}
}
\value{
rawData Raw data loaded into data frame
}
\description{
General function which allows specifying different types of input data
including "proteios", "maxquantpep" (peptide output from MaxQuant) and 
"maxquantprot" (protein output from MaxQuant) formats.
}
\examples{
data_path <- system.file(package="NormalyzerDE", "extdata", "tiny_data.tsv")
df <- loadData(data_path)
}
