% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparsers.R
\name{maxQuantInputOptions}
\alias{maxQuantInputOptions}
\title{Create input options for MaxQuant matrices}
\usage{
maxQuantInputOptions(sep = "\\t")
}
\arguments{
\item{sep}{Field separator used when reading the data matrix.}
}
\value{
A list suitable to pass as \code{inputOptions}.
}
\description{
Helper to construct an \code{inputOptions} list for \code{inputFormat =
"maxquantpep"} and \code{"maxquantprot"}. Currently this supports selecting
the input delimiter.
}
\examples{
maxQuantInputOptions(sep = "\t")
}
