% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normMethods.R
\name{meanNormalization}
\alias{meanNormalization}
\title{Normalization adjusting towards sample mean intensity}
\usage{
meanNormalization(rawMatrix)
}
\arguments{
\item{rawMatrix}{Target matrix to be normalized}
}
\value{
Normalized and log-transformed matrix
}
\description{
Normalization adjusting towards sample mean intensity
}
\examples{
data(example_data_only_values)
normMatrix <- meanNormalization(example_data_only_values)
}
