% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normMethods.R
\name{medianNormalization}
\alias{medianNormalization}
\title{Normalization adjusting towards sample median intensity}
\usage{
medianNormalization(rawMatrix)
}
\arguments{
\item{rawMatrix}{Target matrix to be normalized}
}
\value{
Normalized and log-transformed matrix
}
\description{
Normalization adjusting towards sample median intensity
}
\examples{
data(example_data_only_values)
normMatrix <- medianNormalization(example_data_only_values)
}
