\name{Normalyzer-package}
\alias{Normalyzer-package}
\alias{Normalyzer}
\docType{package}
\title{
Normalyzer
}
\description{
Normalyzer : a tool for rapid evaluation of normalization methods for omics datasets
}
\details{
\tabular{ll}{
Package: \tab Normalyzer\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-04-22\cr
License: \tab GPL-3\cr
}

}
\author{
Aakash Chawade
Maintainer: Aakash Chawade <aakash.chawade@immun.lth.se>
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
