% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerDataset.R
\docType{class}
\name{NormalyzerDataset-class}
\alias{NormalyzerDataset-class}
\alias{NormalyzerDataset}
\title{Represents raw input data together with basic annotation information}
\description{
Takes a job name, a data matrix, a design matrix as well as specification
of the group and sample columns in the design matrix. Provides the
basic representation of a dataset in the NormalyzerDE normalization part.
}
\section{Slots}{

\describe{
\item{\code{jobName}}{Name of the job represented by the dataset.}

\item{\code{rawData}}{Matrix with raw values.}

\item{\code{sampleNameCol}}{Name column for sample.}

\item{\code{groupNameCol}}{Name column for groups.}

\item{\code{designMatrix}}{Data frame containing design.}

\item{\code{sampleNames}}{Vector containing sample names.}

\item{\code{filterrawdata}}{Reduced raw data matrix where low abundance rows are 
removed}

\item{\code{sampleReplicateGroups}}{Vector with sample replicate information}

\item{\code{samplesGroupsWithReplicates}}{Vector with replicated sample replicate information}

\item{\code{colsum}}{Vector with sum of values for each column}

\item{\code{medofdata}}{Vector with median values for each column}

\item{\code{meanofdata}}{Vector with mean values for each column}

\item{\code{annotationValues}}{Annotation part of original dataframe.}

\item{\code{retentionTimes}}{Vector of retention time values.}

\item{\code{singleReplicateRun}}{Conditional whether run is single replicate.}
}}

