% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerEvaluationResults.R
\docType{class}
\name{NormalyzerEvaluationResults-class}
\alias{NormalyzerEvaluationResults-class}
\alias{NormalyzerEvaluationResults}
\title{S4 class to represent normalization evaluations}
\description{
S4 class to represent normalization evaluations
}
\section{Slots}{

\describe{
\item{\code{avgcvmem}}{Average coefficient of variance per method}

\item{\code{avgcvmempdiff}}{Percentage difference of mean coefficient of variance
compared to log2-transformed data}

\item{\code{featureCVPerMethod}}{CV calculated per feature and normalization method.}

\item{\code{avgmadmem}}{Average median absolute deviation}

\item{\code{avgmadmempdiff}}{Percentage difference of median absolute deviation
compared to log2-transformed data}

\item{\code{avgvarmem}}{Average variance per method}

\item{\code{avgvarmempdiff}}{Percentage difference of mean variance compared
to log2-transformed data}

\item{\code{lowVarFeaturesCVs}}{List of 5% least variable entries based on ANOVA
for log2-transformed data}

\item{\code{lowVarFeaturesCVsPercDiff}}{Coefficient of variance for least variable
entries}

\item{\code{anfdr}}{ANOVA FDR values}

\item{\code{anovaFDRWithNA}}{ANOVA FDR with NA when not applicable}

\item{\code{anovaP}}{ANOVA calculated p-values}

\item{\code{repCorPear}}{Within group Pearson correlations}

\item{\code{repCorSpear}}{Within group Spearman correlations}
}}

