% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerStatistics.R
\docType{class}
\name{NormalyzerStatistics-class}
\alias{NormalyzerStatistics-class}
\alias{NormalyzerStatistics}
\title{Statistical dataset representation}
\description{
Statistical dataset representation
}
\section{Slots}{

\describe{
\item{\code{annotMat}}{Matrix containing annotation information}

\item{\code{dataMat}}{Matrix containing (normalized) expression data}

\item{\code{filteredDataMat}}{Filtered matrix with low-count rows removed}

\item{\code{designDf}}{Data frame containing design conditions}

\item{\code{filteringContrast}}{Vector showing which entries are filtered 
(due to low count)}

\item{\code{pairwiseCompsP}}{List with P-values for pairwise comparisons}

\item{\code{pairwiseCompsFdr}}{List with FDR-values for pairwise comparisons}

\item{\code{pairwiseCompsAve}}{List with average expression values}

\item{\code{pairwiseCompsFold}}{List with log2 fold-change values for pairwise 
comparisons}
}}

