% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normMethods.R
\name{normMethods}
\alias{normMethods}
\title{Perform normalizations on Normalyzer dataset}
\usage{
normMethods(nds, currentjob, jobdir)
}
\arguments{
\item{nds}{Normalyzer dataset object.}

\item{currentjob}{Name of the ongoing processing run.}

\item{jobdir}{Path to the output directory}
}
\value{
Returns Normalyzer results object with performed analyzes assigned
 as attributes
}
\description{
Perform normalizations on Normalyzer dataset
}
\examples{
normMethods(my_normalyzer_data_object, "my_run", outputDir="my_run_out_dir")
}
