% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normMethods.R
\name{performCyclicLoessNormalization}
\alias{performCyclicLoessNormalization}
\title{Cyclic Loess normalization
Local regression based on k-nearest-neighbor model}
\usage{
performCyclicLoessNormalization(rawMatrix)
}
\arguments{
\item{rawMatrix}{Target matrix to be normalized}
}
\value{
Normalized matrix
}
\description{
Cyclic Loess normalization
Local regression based on k-nearest-neighbor model
}
\examples{
data(example_data_only_values)
normMatrix <- performCyclicLoessNormalization(example_data_only_values)
}
