% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normMethods.R
\name{performGlobalRLRNormalization}
\alias{performGlobalRLRNormalization}
\title{Global linear regression normalization}
\usage{
performGlobalRLRNormalization(rawMatrix)
}
\arguments{
\item{rawMatrix}{Target matrix to be normalized}
}
\value{
Normalized matrix
}
\description{
Global linear regression normalization
}
\examples{
data(example_data_only_values)
normMatrix <- performGlobalRLRNormalization(example_data_only_values)
}
