% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normMethods.R
\name{performNoNormalization}
\alias{performNoNormalization}
\title{Do no normalization (For debugging purposes)}
\usage{
performNoNormalization(rawMatrix)
}
\arguments{
\item{rawMatrix}{Target matrix to be normalized}
}
\value{
Normalized matrix
}
\description{
Do no normalization (For debugging purposes)
}
\keyword{internal}
