% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerResults.R
\docType{methods}
\name{performNormalizations}
\alias{performNormalizations}
\alias{performNormalizations,NormalyzerResults-method}
\title{Main function for executing normalizations}
\usage{
performNormalizations(nr, forceAll, rtNorm, rtWindow, quiet)

\S4method{performNormalizations}{NormalyzerResults}(nr, forceAll = FALSE,
  rtNorm = FALSE, rtWindow = 0.1, quiet = FALSE)
}
\arguments{
\item{nr}{Normalyzer results object.}

\item{forceAll}{Ignore dataset size limits and run all normalizations
(only meant for testing purposes)}

\item{rtNorm}{Perform retention time based normalizations}

\item{rtWindow}{Retention time window size.}
}
\value{
None
}
\description{
Main function for executing normalizations
}
\keyword{internal}
