% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerResults.R
\docType{methods}
\name{performRTNormalizations}
\alias{performRTNormalizations}
\alias{performRTNormalizations,NormalyzerResults-method}
\title{Perform retention time normalizations}
\usage{
performRTNormalizations(nr, stepSizeMinutes, overlapWindows)

\S4method{performRTNormalizations}{NormalyzerResults}(nr, stepSizeMinutes,
  overlapWindows = FALSE)
}
\arguments{
\item{nr}{Results object.}

\item{stepSizeMinutes}{Size of normalization windows in minutes retention time.}

\item{overlapWindows}{Number of overlapping normalization windows.}
}
\value{
None
}
\description{
Perform retention time normalizations
}
\keyword{internal}
