% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normMethods.R
\name{performSMADNormalization}
\alias{performSMADNormalization}
\title{Median absolute deviation normalization
Normalization subtracts the median and divides the data by the
median absolute deviation (MAD).}
\usage{
performSMADNormalization(rawMatrix)
}
\arguments{
\item{rawMatrix}{Target matrix to be normalized}
}
\value{
Normalized matrix
}
\description{
Median absolute deviation normalization
Normalization subtracts the median and divides the data by the
median absolute deviation (MAD).
}
\examples{
data(example_data_only_values_small)
normMatrix <- performSMADNormalization(example_data_only_values)
}
