% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{plotBoxPlot}
\alias{plotBoxPlot}
\title{Boxplots showing distribution of values after different normalizations}
\usage{
plotBoxPlot(nr, currentLayout, pageno)
}
\arguments{
\item{nr}{Normalyzer results object.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}
}
\value{
None
}
\description{
Boxplots showing distribution of values after different normalizations
}
\keyword{internal}
