% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateStatistics.R
\name{plotComparisonVenns}
\alias{plotComparisonVenns}
\title{If multiple comparisons - Show overlap in Venn diagrams}
\usage{
plotComparisonVenns(
  nst,
  jobName,
  currentLayout,
  pageno,
  sigThres = 0.1,
  sigThresType = "fdr",
  log2FoldThres = 0,
  maxContrasts = 4
)
}
\arguments{
\item{nst}{NormalyzerDE statistics object.}

\item{jobName}{Name of processing run.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}

\item{sigThres}{Cutoff value for significance theshold}

\item{sigThresType}{Type of significance cutoff}

\item{log2FoldThres}{Log2-fold based cutoff threshold}

\item{maxContrasts}{Maximum contrasts to show pairwise comparisons for}
}
\value{
None
}
\description{
If multiple comparisons - Show overlap in Venn diagrams
}
\keyword{internal}
