% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateStatistics.R
\name{plotContrastPCA}
\alias{plotContrastPCA}
\title{Show in a PCA plot what samples are compared in statistical contrast
This is useful to understand what conditions are compared and for checking
for outliers in the contrast}
\usage{
plotContrastPCA(nst, jobName, currentLayout, pageno, pcs = c(1, 2))
}
\arguments{
\item{nst}{NormalyzerDE statistics object.}

\item{jobName}{Name of processing run.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}

\item{pcs}{Principal components to show.}
}
\value{
None
}
\description{
Show in a PCA plot what samples are compared in statistical contrast
This is useful to understand what conditions are compared and for checking
for outliers in the contrast
}
\keyword{internal}
