% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{plotCVvsIntensity}
\alias{plotCVvsIntensity}
\title{Overview of coefficient of variation compared to intensity for variables
 for different normalization methods}
\usage{
plotCVvsIntensity(nr, currentLayout, pageno)
}
\arguments{
\item{nr}{Normalyzer results object.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}
}
\value{
None
}
\description{
Overview of coefficient of variation compared to intensity for variables
 for different normalization methods
}
\keyword{internal}
