% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{plotDendrograms}
\alias{plotDendrograms}
\title{Visualize dendrogram grouping of samples}
\usage{
plotDendrograms(nr, currentLayout, pageno)
}
\arguments{
\item{nr}{Normalyzer results object.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}
}
\value{
None
}
\description{
Visualize dendrogram grouping of samples
}
\keyword{internal}
