% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{plotDensity}
\alias{plotDensity}
\title{Density plots showing value distributions after normalizations}
\usage{
plotDensity(nr, currentLayout, pageno)
}
\arguments{
\item{nr}{Normalyzer results object.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}
}
\value{
None
}
\description{
Density plots showing value distributions after normalizations
}
\keyword{internal}
