% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{plotDEPlots}
\alias{plotDEPlots}
\title{Visualize number of DE variables for ANOVA and Kruskal Wallis}
\usage{
plotDEPlots(nr, currentLayout, pageno)
}
\arguments{
\item{nr}{Normalyzer results object.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}
}
\value{
None
}
\description{
Visualize number of DE variables for ANOVA and Kruskal Wallis
}
\keyword{internal}
