% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{plotFrontPage}
\alias{plotFrontPage}
\title{Generate first page in output report}
\usage{
plotFrontPage(currentjob, currentFont)
}
\arguments{
\item{currentjob}{Name of current run.}

\item{currentFont}{Font used for output document.}
}
\value{
None
}
\description{
Generate first page in output report
}
\keyword{internal}
