% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{plotMA}
\alias{plotMA}
\title{Produces a page containing expression vs. fold-change figures (MA plots)
The visualized fold is between the first sample in each group and the
average of the replicate to which that sample belongs}
\usage{
plotMA(nr, currentLayout, pageno)
}
\arguments{
\item{nr}{Normalyzer results object.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}
}
\value{
None
}
\description{
Produces a page containing expression vs. fold-change figures (MA plots)
The visualized fold is between the first sample in each group and the
average of the replicate to which that sample belongs
}
\keyword{internal}
