% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{plotMDS}
\alias{plotMDS}
\title{MDS plots showing grouping of samples after normalizations}
\usage{
plotMDS(nr, currentLayout, pageno)
}
\arguments{
\item{nr}{Normalyzer results object.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}
}
\value{
None
}
\description{
MDS plots showing grouping of samples after normalizations
}
\keyword{internal}
