% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{plotMeanSD}
\alias{plotMeanSD}
\title{Visualize standard deviation over (expression?) for different values}
\usage{
plotMeanSD(nr, currentLayout, pageno)
}
\arguments{
\item{nr}{Normalyzer results object.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}
}
\value{
None
}
\description{
Visualize standard deviation over (expression?) for different values
}
\keyword{internal}
