% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{plotPHist}
\alias{plotPHist}
\title{Generate P-histograms for ANOVA calculated after each normalization}
\usage{
plotPHist(nr, currentLayout, pageno)
}
\arguments{
\item{nr}{Normalyzer results object.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}
}
\value{
None
}
\description{
Generate P-histograms for ANOVA calculated after each normalization
}
\keyword{internal}
