% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{plotQQ}
\alias{plotQQ}
\title{Produces page showing QQ-plots for the first sample for each normalization
method. This plot can be used to assess whether the data follows a normal
distribution.}
\usage{
plotQQ(nr, currentLayout, pageno)
}
\arguments{
\item{nr}{Normalyzer results object.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}
}
\value{
None
}
\description{
Produces page showing QQ-plots for the first sample for each normalization
method. This plot can be used to assess whether the data follows a normal
distribution.
}
\keyword{internal}
