% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{plotReplicateVariance}
\alias{plotReplicateVariance}
\title{Generate normalization replicate variance summary}
\usage{
plotReplicateVariance(nr, currentLayout, pageno)
}
\arguments{
\item{nr}{Normalyzer results object.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}
}
\value{
None
}
\description{
Generate normalization replicate variance summary
}
\keyword{internal}
