% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{plotRLE}
\alias{plotRLE}
\title{Boxplots showing relative log expression after normalizations}
\usage{
plotRLE(nr, currentLayout, pageno)
}
\arguments{
\item{nr}{Normalyzer results object.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}
}
\value{
None
}
\description{
Boxplots showing relative log expression after normalizations
}
\keyword{internal}
