% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{plotSampleMappingPage}
\alias{plotSampleMappingPage}
\title{Write page with sample mapping}
\usage{
plotSampleMappingPage(nr, currentFont, currentLayout, currentjob, pageno)
}
\arguments{
\item{nr}{Normalyzer results object.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}
}
\value{
None
}
\description{
Write page with sample mapping
}
\keyword{internal}
