% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{preprocessData}
\alias{preprocessData}
\title{Replace 0 values with NA in input data}
\usage{
preprocessData(dataMatrix)
}
\arguments{
\item{dataMatrix}{Matrix with raw data.}
}
\value{
Parsed rawdata where 0 values are replaced with NA
}
\description{
Replace 0 values with NA in input data
}
\keyword{internal}
