% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMeta.R
\name{printMeta}
\alias{printMeta}
\title{Print meta information for Normalyzer plot page
! Needs refactoring to reduce redundancy in code
! Needs double check of functionality}
\usage{
printMeta(plotname, pageno, jobname, currentLayout)
}
\arguments{
\item{plotname}{Name of current plot.}

\item{pageno}{Current page number.}

\item{jobname}{Name of ongoing job.}

\item{currentLayout}{Custom viewport layout.}
}
\value{
None
}
\description{
Print meta information for Normalyzer plot page
! Needs refactoring to reduce redundancy in code
! Needs double check of functionality
}
\keyword{internal}
