% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printPlots.R
\name{printPlots}
\alias{printPlots}
\title{Generate PDF grid page filling it with provided list of plots}
\usage{
printPlots(plotlist, plotname, pageno, jobname, currentLayout)
}
\arguments{
\item{plotlist}{List of target plots to display.}

\item{plotname}{List of names corresponding to the provided plot list.}

\item{pageno}{Current page number.}

\item{jobname}{Name of ongoing job.}

\item{currentLayout}{Custom viewport layout.}
}
\value{
None
}
\description{
Generate PDF grid page filling it with provided list of plots
}
\keyword{internal}
