% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparsers.R
\name{proteiosInputOptions}
\alias{proteiosInputOptions}
\title{Create input options for Proteios matrices}
\usage{
proteiosInputOptions(sep = "\\t")
}
\arguments{
\item{sep}{Field separator used when reading the data matrix.}
}
\value{
A list suitable to pass as \code{inputOptions}.
}
\description{
Helper to construct an \code{inputOptions} list for
\code{inputFormat = "proteios"}. Currently this supports selecting the input
delimiter.
}
\examples{
proteiosInputOptions(sep = ";")
}
