% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateStatistics.R
\name{reduceDesignTechRep}
\alias{reduceDesignTechRep}
\title{Remove technical replicates from design matrix.}
\usage{
reduceDesignTechRep(designMat, techRepGroups)
}
\arguments{
\item{designMat}{NormalyzerDE design matrix}

\item{techRepGroups}{Technical replicates vector}
}
\value{
collDesignDf Reduced design matrix
}
\description{
Technical replicates are specified as duplicate strings.
The first sample name corresponding for each technical replicate group
is retained.
}
\examples{
designDf <- data.frame(
  sample=c("a1", "a2", "b1", "b2", "c1", "c2", "d1", "d2"),
  group=c(rep("A", 4), rep("B", 4)),
  techrep=c("a", "a", "b", "b", "c", "c", "d", "d")
)
statObj <- reduceDesignTechRep(designDf, designDf$techrep)
}
