% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regression_test_nr}
\alias{regression_test_nr}
\title{NormalyzerDE results object used internally for testing}
\format{An instance of the class NormalyzerResults}
\usage{
regression_test_nr
}
\description{
NormalyzerDE results object used internally for testing
}
\keyword{internal}
