% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerDataset.R
\docType{methods}
\name{setupBasicValues}
\alias{setupBasicValues}
\alias{setupBasicValues,NormalyzerDataset-method}
\title{Setup basic values for dataset}
\usage{
setupBasicValues(nds)

\S4method{setupBasicValues}{NormalyzerDataset}(nds)
}
\arguments{
\item{nds}{Normalyzer dataset}
}
\value{
Updated dataset object
}
\description{
Setup basic values for dataset
}
\keyword{internal}
