% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerDataset.R
\docType{methods}
\name{setupFilterRawData}
\alias{setupFilterRawData}
\alias{setupFilterRawData,NormalyzerDataset-method}
\title{Setup filtered raw data slot}
\usage{
setupFilterRawData(nds)

\S4method{setupFilterRawData}{NormalyzerDataset}(nds)
}
\arguments{
\item{nds}{Normalyzer dataset.}
}
\value{
None
}
\description{
Setup filtered raw data slot
}
\keyword{internal}
