% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeResults.R
\name{setupNormalizationEvaluationObject}
\alias{setupNormalizationEvaluationObject}
\title{Setup normalization evaluation object}
\usage{
setupNormalizationEvaluationObject(nr, comparisons = NULL,
  categoricalAnova = FALSE, varFilterFrac = NULL)
}
\arguments{
\item{nr}{Normalyzer results object to be evaluated}

\item{comparisons}{Group comparisons to make}

\item{categoricalAnova}{If comparison should be made numerically or groupwise}

\item{varFilterFrac}{Filtering high-variance samples}
}
\value{
Normalization evaluation object
}
\description{
Setup normalization evaluation object
}
\keyword{internal}
