% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{setupPlotting}
\alias{setupPlotting}
\title{Setup PDF report settings}
\usage{
setupPlotting(currentJob, jobDir, suffix)
}
\arguments{
\item{currentJob}{Name of current run.}

\item{jobDir}{Path to output directory for run.}

\item{suffix}{Text to add to output filename.}
}
\value{
None
}
\description{
Setup PDF report settings
}
\keyword{internal}
