% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{setupRawContrastObject}
\alias{setupRawContrastObject}
\title{Prepare SummarizedExperiment object for statistics data}
\usage{
setupRawContrastObject(
  dataPath,
  designPath,
  sampleColName,
  inputFormat = "default",
  inputOptions = NULL
)
}
\arguments{
\item{dataPath}{Path to raw data matrix}

\item{designPath}{Path to design matrix}

\item{sampleColName}{Name for column in design matrix containing sample names}

\item{inputFormat}{Type of input format for \code{dataPath}. Supports
\code{"default"} and \code{"diann"}.}

\item{inputOptions}{Optional list of input-reader options. Used when
\code{inputFormat="diann"} to select columns and filters (q-value cutoffs,
decoys, min-positive threshold, RT summarization, etc.).}
}
\value{
experimentObj Prepared instance of SummarizedExperiment
}
\description{
Prepare SummarizedExperiment object for statistics data
}
\examples{
data_path <- system.file(package="NormalyzerDE", "extdata", "tiny_data.tsv")
design_path <- system.file(package="NormalyzerDE", "extdata", "tiny_design.tsv")
sumExpObj <- setupRawContrastObject(data_path, design_path, "sample")
}
