% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerDataset.R
\docType{methods}
\name{setupRTColumn}
\alias{setupRTColumn}
\alias{setupRTColumn,NormalyzerDataset-method}
\title{Check for and set up RT column if single -1 annotation present}
\usage{
setupRTColumn(nds, quiet)

\S4method{setupRTColumn}{NormalyzerDataset}(nds, quiet = FALSE)
}
\arguments{
\item{nds}{Normalyzer dataset}
}
\value{
Updated dataset object
}
\description{
Check for and set up RT column if single -1 annotation present
}
\keyword{internal}
