% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setupTestData}
\alias{setupTestData}
\title{Generate a random test dataset with features, sample values and retention times}
\usage{
setupTestData(nSamples, nFeatures, rtMin = 40, rtMax = 80, mean = 20,
  sd = 4)
}
\arguments{
\item{nSamples}{Number of samples}

\item{nFeatures}{Number of features}

\item{rtMin}{Minimum retention time}

\item{rtMax}{Maximum retention time}

\item{mean}{Mean value for sample intensities}

\item{sd}{Standard deviation for sample intensities}
}
\value{
Test dataset
}
\description{
Generate a random test dataset with features, sample values and retention times
}
\examples{
df <- setupTestData(6, 20)
df <- setupTestData(6, 20, mean=15, sd=1)
}
\keyword{internal}
