% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setupTestDesign}
\alias{setupTestDesign}
\title{Generate design matrix}
\usage{
setupTestDesign(nSamples)
}
\arguments{
\item{nSamples}{Number of samples}
}
\value{
Design matrix
}
\description{
Generate design matrix
}
\examples{
df <- setupTestDesign(6)
}
\keyword{internal}
