% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerDataset.R
\docType{methods}
\name{setupValues}
\alias{setupValues}
\alias{setupValues,NormalyzerDataset-method}
\title{Initialize values for dataset}
\usage{
setupValues(nds, quiet)

\S4method{setupValues}{NormalyzerDataset}(nds, quiet = FALSE)
}
\arguments{
\item{nds}{Normalyzer dataset}
}
\value{
None
}
\description{
Initialize values for dataset
}
\keyword{internal}
