% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{validateSampleReplication}
\alias{validateSampleReplication}
\title{Check whether all samples have replicates}
\usage{
validateSampleReplication(
  dataMatrix,
  groups,
  requireReplicates = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{dataMatrix}{Prepared matrix containing expression data.}

\item{groups}{Vector containing condition levels}

\item{requireReplicates}{By default stops processing if not all samples
have replicates}
}
\value{
None
}
\description{
Check whether all samples have replicates
}
\keyword{internal}
