% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{verifyDesignMatrix}
\alias{verifyDesignMatrix}
\title{Verify that design matrix setup matches the data matrix}
\usage{
verifyDesignMatrix(fullMatrix, designMatrix, sampleCol)
}
\arguments{
\item{fullMatrix}{Data frame with input data.}

\item{designMatrix}{Data frame with design setup.}

\item{sampleCol}{Column in design matrix containing sample IDs.}
}
\value{
None
}
\description{
Verify that design matrix setup matches the data matrix
}
\keyword{internal}
