% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{verifyMultipleSamplesPresent}
\alias{verifyMultipleSamplesPresent}
\title{Check whether more than one sample is present}
\usage{
verifyMultipleSamplesPresent(dataMatrix, groups, requireReplicates = TRUE,
  quiet = FALSE)
}
\arguments{
\item{dataMatrix}{Prepared dataframe.}

\item{groups}{Vector containing condition levels}

\item{requireReplicates}{By default stops processing if not all samples
have replicates}
}
\value{
None
}
\description{
Check whether more than one sample is present
}
\keyword{internal}
