% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{verifySummarizedExperiment}
\alias{verifySummarizedExperiment}
\title{Verify a SummarizedExperiment contains matching samples}
\usage{
verifySummarizedExperiment(summarizedExp, sampleCol)
}
\arguments{
\item{summarizedExp}{SummarizedExperiment object to validate.}

\item{sampleCol}{Column in \code{colData} containing sample IDs.}
}
\value{
None
}
\description{
Checks that the sample names (columns) present in a \code{SummarizedExperiment}
object match the sample IDs in its \code{colData}.
}
\keyword{internal}
