% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{verifyValidNumbers}
\alias{verifyValidNumbers}
\title{Verify that input fields conform to the expected formats}
\usage{
verifyValidNumbers(rawDataOnly, groups, noLogTransform = FALSE, quiet = FALSE)
}
\arguments{
\item{rawDataOnly}{Data frame with input data.}

\item{groups}{Condition levels for comparisons.}
}
\value{
None
}
\description{
Verify that input fields conform to the expected formats
}
\keyword{internal}
