% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerStatistics.R
\docType{methods}
\name{calculateContrasts}
\alias{calculateContrasts}
\alias{calculateContrasts,NormalyzerStatistics-method}
\title{Main method for contrast calculation}
\usage{
calculateContrasts(nst, comparisons, condCol, batchCol = NULL,
  splitter = "-", type = "limma")

\S4method{calculateContrasts}{NormalyzerStatistics}(nst, comparisons, condCol,
  batchCol = NULL, splitter = "-", type = "limma")
}
\arguments{
\item{nst}{Results evaluation object.}

\item{comparisons}{String with comparisons for contrasts.}

\item{condCol}{Column name in design matrix containing condition information.}

\item{batchCol}{Column name in design matrix containing batch information.}

\item{splitter}{Character dividing contrast conditions.}

\item{type}{Type of statistical test (Limma or ANOVA).}
}
\value{
nst Statistics object with statistical measures calculated
}
\description{
Main method for contrast calculation
}
