% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerEvaluationResults.R
\docType{methods}
\name{calculateCorrelations}
\alias{calculateCorrelations}
\alias{calculateCorrelations,NormalyzerEvaluationResults-method}
\title{Pearson and Spearman correlation calculations for methods and samples
Calculates internal correlation per condition}
\usage{
calculateCorrelations(ner, nr)

\S4method{calculateCorrelations}{NormalyzerEvaluationResults}(ner, nr)
}
\arguments{
\item{ner}{Normalyzer evaluation object.}

\item{nr}{Normalyzer results object with calculated results.}
}
\value{
ner Normalyzer evaluation object with attached evaluation results.
}
\description{
Pearson and Spearman correlation calculations for methods and samples
Calculates internal correlation per condition
}
\keyword{internal}
