% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeResults.R
\name{calculateFeatureCV}
\alias{calculateFeatureCV}
\title{Calculate CV values for each feature}
\usage{
calculateFeatureCV(methodList)
}
\arguments{
\item{methodList}{List containing normalized matrices.}

\item{sampleReplicateGroups}{Condition header.}
}
\value{
methodFeatureCVMatrix Matrix with feature as rows and normalization
  method as columns
}
\description{
Calculate CV values for each feature
}
\keyword{internal}
