% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerDataset.R
\name{detectSingleReplicate}
\alias{detectSingleReplicate}
\alias{detectSingleReplicate,NormalyzerDataset-method}
\title{Detect single replicate, and assign related logical}
\usage{
detectSingleReplicate(nds, quiet)

\S4method{detectSingleReplicate}{NormalyzerDataset}(nds, quiet = FALSE)
}
\arguments{
\item{nds}{Normalyzer dataset}

\item{quiet}{Don't give non-error output}
}
\value{
bool on whether sample contains only one sample group
}
\description{
Detect single replicate, and assign related logical
}
\keyword{internal}
