% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerDataset.R
\docType{methods}
\name{detectSingletonSample}
\alias{detectSingletonSample}
\alias{detectSingletonSample,NormalyzerDataset-method}
\title{Detect single sample group}
\usage{
detectSingletonSample(nds, quiet)

\S4method{detectSingletonSample}{NormalyzerDataset}(nds, quiet = FALSE)
}
\arguments{
\item{nds}{Normalyzer dataset.}

\item{quiet}{Only print error messages}
}
\value{
None
}
\description{
Detect single sample group
}
\keyword{internal}
