% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data_only_values}
\alias{example_data_only_values}
\title{Small example dataset only containing raw numerical values}
\format{A data frame containing expression data}
\usage{
example_data_only_values
}
\description{
Small example dataset only containing raw numerical values
}
\keyword{datasets}
