% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_design}
\alias{example_design}
\title{Design matrix corresponding to the small example datasets.}
\format{A design matrix corresponding to the dataset "example_data"}
\usage{
example_design
}
\description{
Design matrix corresponding to the small example datasets.
}
\keyword{internal}
