% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_stat_data}
\alias{example_stat_data}
\title{Small example normalized dataset}
\format{A normalized data frame ready for statistical processing}
\usage{
example_stat_data
}
\description{
Small example normalized dataset
}
\keyword{datasets}
