% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateStatistics.R
\name{filterLowRep}
\alias{filterLowRep}
\title{Filter rows with lower than given number of replicates for any condition}
\usage{
filterLowRep(df, groups, leastRep = 2)
}
\arguments{
\item{df}{Dataframe with expression data to filter}

\item{groups}{Condition groups header}

\item{leastRep}{Minimum number of replicates in each group
to retain}
}
\value{
collDesignDf Reduced design matrix
}
\description{
Filter rows with lower than given number of replicates for any condition
}
\keyword{internal}
