% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{generatePlots}
\alias{generatePlots}
\title{Generate full output report plot document}
\usage{
generatePlots(nr, jobdir, plotRows = 3, plotCols = 4)
}
\arguments{
\item{nr}{Normalyzer results object.}

\item{jobdir}{Path to output directory for run.}

\item{plotRows}{Number of plot rows.}

\item{plotCols}{Number of plot columns.}
}
\value{
None
}
\description{
Generate full output report plot document
}
\examples{
data(example_data)
data(example_design)
normObj <- getVerifiedNormalyzerObject("job_name", example_design, example_data)
normResults <- normMethods(normObj)
normResultsWithEval <- analyzeNormalizations(normResults)
outputDir <- tempdir()
generatePlots(normResultsWithEval, outputDir)
}
