% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateStatistics.R
\name{generateStatsReport}
\alias{generateStatsReport}
\title{Generate full output report plot document}
\usage{
generateStatsReport(nst, jobName, jobDir, plotRows = 3, plotCols = 4)
}
\arguments{
\item{nst}{NormalyzerDE statistics object.}

\item{jobName}{Name of processing run.}

\item{jobDir}{Path to output directory.}

\item{plotRows}{Number of plot rows.}

\item{plotCols}{Number of plot columns.}
}
\value{
None
}
\description{
Generate full output report plot document
}
\examples{
data(example_stat_data)
data(example_design)
statObj <- setupStatisticsObject(example_design, example_stat_data)
statObj <- calculateContrasts(statObj, comparisons=c("1-2", "2-3"), condCol="group", type="limma")
outputDir <- tempdir()
generateStatsReport(statObj, "jobName", outputDir)
}
