% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{getLowCountSampleFiltered}
\alias{getLowCountSampleFiltered}
\title{Verify that samples contain at least a lowest number of values}
\usage{
getLowCountSampleFiltered(dataMatrix, groups, threshold = 15,
  stopIfTooFew = TRUE)
}
\arguments{
\item{dataMatrix}{Dataframe with processed input data.}

\item{groups}{Vector containing condition levels.}

\item{threshold}{Lowest number of allowed values in a column.}

\item{stopIfTooFew}{Abort run if lower than threshold number of values in
column}
}
\value{
None
}
\description{
Verify that samples contain at least a lowest number of values
}
\keyword{internal}
