% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRowNAFilterContrast}
\alias{getRowNAFilterContrast}
\title{Get contrast vector (TRUE/FALSE-values) indicating whether both at least
half values are present, and each sample has at least one non-NA value}
\usage{
getRowNAFilterContrast(dataMatrix, replicateHeader, minCount = 1)
}
\arguments{
\item{dataMatrix}{Matrix with expression values for entities in replicate 
samples.}

\item{replicateHeader}{Header showing how samples in matrix are replicated.}

\item{minCount}{Minimum number of required values present in samples.}
}
\value{
Contrast vector
}
\description{
Get contrast vector (TRUE/FALSE-values) indicating whether both at least
half values are present, and each sample has at least one non-NA value
}
\keyword{internal}
