% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higherOrderNormMethods.R
\name{getWidenedRTRange}
\alias{getWidenedRTRange}
\title{Pick datapoints before and after window until a minimum number is reached}
\usage{
getWidenedRTRange(rtStart, rtEnd, minimumDatapoints, retentionTimes)
}
\arguments{
\item{rtStart}{Original retention time start point}

\item{rtEnd}{Original retention time end point}

\item{minimumDatapoints}{Required number of datapoints to fullfil}

\item{retentionTimes}{Vector with all retention times}
}
\value{
Vector with start and end of new RT range
}
\description{
Pick datapoints before and after window until a minimum number is reached
}
\keyword{internal}
