% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{loadData}
\alias{loadData}
\title{Load raw data into dataframe}
\usage{
loadData(dataPath, inputFormat = "default", zeroToNA = FALSE)
}
\arguments{
\item{dataPath}{File path to design matrix.}

\item{inputFormat}{If input is given in standard NormalyzerDE format, Proteios format
or in MaxQuant protein or peptide format}

\item{zeroToNA}{Automatically convert zeroes to NA values}
}
\value{
rawData Raw data loaded into data frame
}
\description{
Load raw data into dataframe
}
\examples{
\dontrun{
df <- loadData("data.tsv")
}
}
