% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{loadDesign}
\alias{loadDesign}
\title{Load raw design into dataframe}
\usage{
loadDesign(designPath, sampleCol = "sample", groupCol = "group")
}
\arguments{
\item{designPath}{File path to design matrix.}

\item{sampleCol}{Column name for column containing sample names.}

\item{groupCol}{Column name for column containing condition levels.}
}
\value{
designMatrix Design data loaded into data frame
}
\description{
Load raw design into dataframe
}
\examples{
\dontrun{
df <- loadDesign("design.tsv")
}
}
