% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{loadRawDataFromFile}
\alias{loadRawDataFromFile}
\title{Try reading raw Normalyzer matrix from provided filepath}
\usage{
loadRawDataFromFile(inputPath)
}
\arguments{
\item{inputPath}{Path to Normalyzer data.}
}
\value{
Table containing raw data from input file.
}
\description{
Try reading raw Normalyzer matrix from provided filepath
}
\keyword{internal}
