% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normMethods.R
\name{meanNormalization}
\alias{meanNormalization}
\title{Intensity of each variable in a given sample is divided by the mean of sum of
intensities of all variables in the sample and then multiplied by the mean
of sum of intensities of all variables in all samples. The normalized data
is then transformed to log2.}
\usage{
meanNormalization(rawMatrix, noLogTransform = FALSE)
}
\arguments{
\item{rawMatrix}{Target matrix to be normalized}

\item{noLogTransform}{Assumes no need for log transformation}
}
\value{
Normalized and log-transformed matrix
}
\description{
Intensity of each variable in a given sample is divided by the mean of sum of
intensities of all variables in the sample and then multiplied by the mean
of sum of intensities of all variables in all samples. The normalized data
is then transformed to log2.
}
\examples{
data(example_data_only_values_small)
normMatrix <- meanNormalization(example_data_only_values)
}
