% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalyzer-str.R
\name{normalyzer}
\alias{normalyzer}
\title{Normalyzer pipeline entry point}
\usage{
normalyzer(datafile, jobName, outputDir = NULL)
}
\arguments{
\item{datafile}{CSV delimited input file containing raw counts and 
replicate information in header.}

\item{jobName}{Give the current run a name.}

\item{outputDir}{Specify an output directory for generated files.
Defaults to current working directory.}
}
\description{
Normalyzer pipeline entry point
}
\examples{
normalyzer("path/to/input.csv", "my_test_run")
normalyzer("path/to/input.csv", "my_test_run", outputDir="my_output")
}
