% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerDE.R
\name{normalyzer}
\alias{normalyzer}
\title{Normalyzer pipeline entry point}
\usage{
normalyzer(jobName, designPath, dataPath, outputDir = ".",
  forceAllMethods = FALSE, omitLowAbundSamples = FALSE,
  sampleAbundThres = 15, requireReplicates = TRUE,
  normalizeRetentionTime = TRUE, plotRows = 3, plotCols = 4,
  zeroToNA = FALSE, sampleColName = "sample", groupColName = "group",
  inputFormat = "default", skipAnalysis = FALSE, quiet = FALSE,
  rtStepSizeMinutes = 1, rtWindowMinCount = 100, rtWindowShifts = 1,
  rtWindowMergeMethod = "mean")
}
\arguments{
\item{jobName}{Give the current run a name.}

\item{designPath}{Path to file containing design matrix.}

\item{dataPath}{Specify an output directory for generated files.
Defaults to current working directory.}

\item{outputDir}{Directory where results folder is created.}

\item{forceAllMethods}{Debugging function. Run all normalizations even if
they aren't in the recommended range of number of values}

\item{omitLowAbundSamples}{Automatically remove samples with fewer non-NA
values compared to threshold given by sampleAbundThres. 
Will otherwise stop with error message if such sample is encountered.}

\item{sampleAbundThres}{Threshold for omitting low-abundant
samples. Is by default set to 15.}

\item{requireReplicates}{Require multiple samples per condition to pass input 
validation.}

\item{normalizeRetentionTime}{Perform normalizations over retention time.}

\item{plotRows}{Number of plot-rows in output documentation.}

\item{plotCols}{Number of plot-columns in output documentation.}

\item{zeroToNA}{Convert zero values to NA.}

\item{sampleColName}{Column name in design matrix containing sample IDs.}

\item{groupColName}{Column name in design matrix containing condition IDs.}

\item{inputFormat}{Type of input format.}

\item{skipAnalysis}{Only perform normalization steps.}

\item{quiet}{Omit status messages printed during run}

\item{rtStepSizeMinutes}{Retention time normalization window size.}

\item{rtWindowMinCount}{Minimum number of datapoints in each retention-time
segment.}

\item{rtWindowShifts}{Number of layered retention time normalized windows.}

\item{rtWindowMergeMethod}{Merge approach for layered retention time windows.}
}
\value{
None
}
\description{
Normalyzer pipeline entry point
}
\examples{
\dontrun{
normalyzer(
    "data.tsv", 
    "my_jobname", 
    designMatrix="design.tsv", 
    outputDir="path/to/output")
normalyzer(
    "data.tsv", 
    "my_jobname", 
    designMatrix="design.tsv", 
    outputDir="path/to/output", 
    normalizeRetentionTime=TRUE, 
    retentionTimeWindow=2)
normalyzer(
    "data.tsv", 
    "my_jobname", 
    designMatrix="design.tsv", 
    outputDir="path/to/output", 
    inputFormat="maxquantprot")
}
}
