% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerDE.R
\name{normalyzerDE}
\alias{normalyzerDE}
\title{Normalyzer differential expression}
\usage{
normalyzerDE(jobName, designPath, dataPath, comparisons, outputDir = ".",
  logTrans = FALSE, type = "limma", sampleCol = "sample",
  condCol = "group", batchCol = NULL, techRepCol = NULL,
  leastRepCount = 1, quiet = FALSE)
}
\arguments{
\item{jobName}{Name of job}

\item{designPath}{File path to design matrix}

\item{dataPath}{File path to normalized matrix}

\item{comparisons}{Character vector containing target contrasts. 
If comparing condA with condB, then the vector would be c("condA-condB")}

\item{outputDir}{Path to output directory}

\item{logTrans}{Log transform the input (needed if providing non-logged 
input)}

\item{type}{Type of statistical comparison, "limma" or "welch"}

\item{sampleCol}{Design matrix column header for column containing sample IDs}

\item{condCol}{Design matrix column header for column containing sample 
conditions}

\item{batchCol}{Provide an optional column for inclusion of possible batch 
variance in the model}

\item{techRepCol}{Design matrix column header for column containing technical 
replicates}

\item{leastRepCount}{Minimum required replicate count}

\item{quiet}{Omit status messages printed during run}
}
\value{
None
}
\description{
Normalyzer differential expression
}
\examples{
\dontrun{
normalyzerDE(
  "results/normalized.tsv", "design.tsv", "my_jobname", 
  c("1-2", "1-4"),  outputDir="path/to/output")
}
}
