% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normMethods.R
\name{normMethods}
\alias{normMethods}
\title{Perform normalizations on Normalyzer dataset}
\usage{
normMethods(nds, forceAll = FALSE, normalizeRetentionTime = TRUE,
  retentionTimeWindow = 1, quiet = FALSE)
}
\arguments{
\item{nds}{Normalyzer dataset object.}

\item{forceAll}{Force all methods to run despite not qualifying for thresholds.}

\item{normalizeRetentionTime}{Perform retention time based normalization methods.}

\item{retentionTimeWindow}{Default window size for retention times.}

\item{quiet}{Prevent diagnostic output}
}
\value{
Returns Normalyzer results object with performed analyzes assigned
 as attributes
}
\description{
Perform normalizations on Normalyzer dataset
}
\examples{
data(example_data)
data(example_design)
normObj <- getVerifiedNormalyzerObject("job_name", example_design, example_data)
normResults <- normMethods(normObj)
}
