% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normMethods.R
\name{performCyclicLoessNormalization}
\alias{performCyclicLoessNormalization}
\title{Cyclic Loess normalization}
\usage{
performCyclicLoessNormalization(rawMatrix)
}
\arguments{
\item{rawMatrix}{Target matrix to be normalized}
}
\value{
Normalized matrix
}
\description{
Log2 transformed data is normalized by Loess method using the function
"normalizeCyclicLoess". Further information is available for the function
"normalizeCyclicLoess" in the Limma package.
}
\examples{
data(example_data_only_values_small)
normMatrix <- performCyclicLoessNormalization(example_data_only_values)
}
