% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normMethods.R
\name{performQuantileNormalization}
\alias{performQuantileNormalization}
\title{Perform quantile normalization, where overall distribution is assumed
to be the same, and using a 'average' distribution as measure to normalize
again}
\usage{
performQuantileNormalization(rawMatrix)
}
\arguments{
\item{rawMatrix}{Target matrix to be normalized}
}
\value{
Normalized matrix
}
\description{
Perform quantile normalization, where overall distribution is assumed
to be the same, and using a 'average' distribution as measure to normalize
again
}
\examples{
data(example_data_only_values)
normMatrix <- performQuantileNormalization(example_data_only_values)
}
