% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerResults.R
\docType{methods}
\name{performRTNormalizations}
\alias{performRTNormalizations}
\alias{performRTNormalizations,NormalyzerResults-method}
\title{Perform retention time normalizations}
\usage{
performRTNormalizations(nr, stepSizeMinutes, minWindowSize, windowShifts,
  mergeMethod)

\S4method{performRTNormalizations}{NormalyzerResults}(nr, stepSizeMinutes,
  minWindowSize, windowShifts, mergeMethod)
}
\arguments{
\item{nr}{Results object.}

\item{stepSizeMinutes}{Size of normalization windows in minutes retention time.}

\item{minWindowSize}{Minimum number of datapoints within each retention time window.}

\item{windowShifts}{Number of overlapping retention-time segmented normalizations.}

\item{mergeMethod}{Type of merge approach for overlapping windows.}
}
\value{
nr NormalyzerDE results object
}
\description{
Perform retention time normalizations
}
\keyword{internal}
