% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normMethods.R
\name{performVSNNormalization}
\alias{performVSNNormalization}
\title{Log2 transformed data is normalized using the function "justvsn" from the
VSN package.}
\usage{
performVSNNormalization(rawMatrix)
}
\arguments{
\item{rawMatrix}{Target matrix to be normalized}
}
\value{
Normalized matrix
}
\description{
The VSN (Variance Stabilizing Normalization) attempts to transform the data 
in such a way that the variance remains nearly constant over the intensity 
spectrum
}
\examples{
data(example_data_only_values)
normMatrix <- performVSNNormalization(example_data_only_values)
}
