% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normMethods.R
\name{performVSNNormalization}
\alias{performVSNNormalization}
\title{Perform Variance Stabilizing Normalization}
\usage{
performVSNNormalization(rawMatrix)
}
\arguments{
\item{rawMatrix}{Target matrix to be normalized}
}
\value{
Normalized matrix
}
\description{
Perform Variance Stabilizing Normalization
}
\examples{
data(example_data_only_values)
normMatrix <- performVSNNormalization(example_data_only_values)
}
