% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateStatistics.R
\name{plotContrastPHists}
\alias{plotContrastPHists}
\title{Takes an NormalyzerStatistics instance and generates and prints a p-value
histogram for each onto the viewport}
\usage{
plotContrastPHists(nst, jobName, currentLayout, pageno)
}
\arguments{
\item{nst}{NormalyzerDE statistics object.}

\item{jobName}{Name of processing run.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}
}
\value{
None
}
\description{
Takes an NormalyzerStatistics instance and generates and prints a p-value
histogram for each onto the viewport
}
\keyword{internal}
