% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{plotCorrelation}
\alias{plotCorrelation}
\title{Visualize within-replicates correlations}
\usage{
plotCorrelation(nr, currentLayout, pageno)
}
\arguments{
\item{nr}{Normalyzer results object.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}
}
\value{
None
}
\description{
Visualize within-replicates correlations
}
\keyword{internal}
