% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{plotCVvsIntensity}
\alias{plotCVvsIntensity}
\title{Plots page displaying coefficient of variance (CV) against raw intensity
for features across the performed normalizations}
\usage{
plotCVvsIntensity(nr, currentLayout, pageno)
}
\arguments{
\item{nr}{Normalyzer results object.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}
}
\value{
None
}
\description{
Plots page displaying coefficient of variance (CV) against raw intensity
for features across the performed normalizations
}
\keyword{internal}
