% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{plotMA}
\alias{plotMA}
\title{Expression vs. fold-change for variables for each normalization method}
\usage{
plotMA(nr, currentLayout, pageno)
}
\arguments{
\item{nr}{Normalyzer results object.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}
}
\value{
None
}
\description{
Expression vs. fold-change for variables for each normalization method
}
\keyword{internal}
