% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{plotQQ}
\alias{plotQQ}
\title{QQ-plots for variable values}
\usage{
plotQQ(nr, currentLayout, pageno)
}
\arguments{
\item{nr}{Normalyzer results object.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}
}
\value{
None
}
\description{
QQ-plots for variable values
}
\keyword{internal}
