% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{plotReplicateVarAndStableVariables}
\alias{plotReplicateVarAndStableVariables}
\title{Write figures displaying pooled coefficient of variance, median absolute 
deviation and pooled estimate of variance percentage compared to 
log2-transformed and stable variables plot displaying CV of stable variables
against pooled CV measure. The stable variables are calculated by an 
ANOVA comparison across sample conditions and selecting features with the 
least clear difference.}
\usage{
plotReplicateVarAndStableVariables(nr, currentLayout, pageno)
}
\arguments{
\item{nr}{Normalyzer results object.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}
}
\value{
None
}
\description{
Write figures displaying pooled coefficient of variance, median absolute 
deviation and pooled estimate of variance percentage compared to 
log2-transformed and stable variables plot displaying CV of stable variables
against pooled CV measure. The stable variables are calculated by an 
ANOVA comparison across sample conditions and selecting features with the 
least clear difference.
}
\keyword{internal}
