% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{plotReplicateVarAndStableVariables}
\alias{plotReplicateVarAndStableVariables}
\title{Generate replicate variance overview for normalization methods and CV
plot for stable variables}
\usage{
plotReplicateVarAndStableVariables(nr, currentLayout, pageno)
}
\arguments{
\item{nr}{Normalyzer results object.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}
}
\value{
None
}
\description{
Generate replicate variance overview for normalization methods and CV
plot for stable variables
}
\keyword{internal}
