% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{plotSampleOutlierSummary}
\alias{plotSampleOutlierSummary}
\title{Generate sample summary of intensities, missing values and MDS plot}
\usage{
plotSampleOutlierSummary(nr, currentLayout, pageno)
}
\arguments{
\item{nr}{Normalyzer results object.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}
}
\value{
None
}
\description{
Generate sample summary of intensities, missing values and MDS plot
}
\keyword{internal}
