% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{plotSampleOutlierSummary}
\alias{plotSampleOutlierSummary}
\title{Write page containing sample summary of intensities, missing values and 
MDS plot to the viewport}
\usage{
plotSampleOutlierSummary(nr, currentLayout, pageno)
}
\arguments{
\item{nr}{Normalyzer results object.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}
}
\value{
None
}
\description{
Write page containing sample summary of intensities, missing values and 
MDS plot to the viewport
}
\keyword{internal}
