% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{plotScatter}
\alias{plotScatter}
\title{Scatter plot comparing replicate expression between samples}
\usage{
plotScatter(nr, currentLayout, pageno)
}
\arguments{
\item{nr}{Normalyzer results object.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}
}
\value{
None
}
\description{
Scatter plot comparing replicate expression between samples
}
\keyword{internal}
