% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{plotScatter}
\alias{plotScatter}
\title{Produces page containing scatter plot plotting the first two samples from 
each dataset against each other for each normalization method}
\usage{
plotScatter(nr, currentLayout, pageno)
}
\arguments{
\item{nr}{Normalyzer results object.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}
}
\value{
None
}
\description{
Produces page containing scatter plot plotting the first two samples from 
each dataset against each other for each normalization method
}
\keyword{internal}
