% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateStatistics.R
\name{plotSigScatter}
\alias{plotSigScatter}
\title{Takes an NormalyzerStatistics instance and generates and prints a 
vulcano plot}
\usage{
plotSigScatter(nst, jobName, currentLayout, pageno, type = "Vulcano",
  sigThres = 0.1)
}
\arguments{
\item{nst}{NormalyzerDE statistics object.}

\item{jobName}{Name of processing run.}

\item{currentLayout}{Layout used for document.}

\item{pageno}{Current page number.}

\item{type}{Specify whether to plot 'Vulcano' or 'MA'.}

\item{sigThres}{FDR threshold for DE coloring.}
}
\value{
None
}
\description{
Takes an NormalyzerStatistics instance and generates and prints a 
vulcano plot
}
\keyword{internal}
