% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{preprocessData}
\alias{preprocessData}
\title{Replace empty values (0 or empty field) with NA in input data}
\usage{
preprocessData(dataMatrix, quiet = FALSE)
}
\arguments{
\item{dataMatrix}{Matrix with raw data.}

\item{quiet}{Don't show diagnostic messages}
}
\value{
Parsed rawdata where 0 values are replaced with NA
}
\description{
Replace empty values (0 or empty field) with NA in input data
}
\keyword{internal}
