% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateStatistics.R
\name{reduceTechnicalReplicates}
\alias{reduceTechnicalReplicates}
\title{Remove technical replicates from data matrix}
\usage{
reduceTechnicalReplicates(dataMat, techRepGroups)
}
\arguments{
\item{dataMat}{NormalyzerDE data matrix}

\item{techRepGroups}{Technical replicates vector}
}
\value{
collDataMat Reduced data matrix
}
\description{
Collapses sample values into their average. If only one value is present
due to NA-values in other technical replicates, then that value is used.
}
\examples{
techRep <- c("a", "a", "b", "b", "c", "c", "d", "d")
testData <- data.frame(
    c(1,1,1), 
    c(1,2,1), 
    c(3,3,3), 
    c(5,3,3), 
    c(5,5,4), 
    c(5,5,5), 
    c(7,7,7), 
    c(7,9,7))
colnames(testData) <- c("a1", "a2", "b1", "b2", "c1", "c2", "d1", "d2")
statObj <- reduceTechnicalReplicates(testData, techRep)
}
