% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{setupRawContrastObject}
\alias{setupRawContrastObject}
\title{Prepare SummarizedExperiment object for statistics data}
\usage{
setupRawContrastObject(dataPath, designPath, sampleColName)
}
\arguments{
\item{dataPath}{Path to raw data matrix}

\item{designPath}{Path to design matrix}

\item{sampleColName}{Name for column in design matrix containing sample names}
}
\value{
experimentObj Prepared instance of SummarizedExperiment
}
\description{
Prepare SummarizedExperiment object for statistics data
}
\examples{
\dontrun{
sumExpObj <- setupRawContrastObject("data.tsv", "design.tsv", "sample")
}
}
