% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateStatistics.R
\name{setupStatisticsObject}
\alias{setupStatisticsObject}
\title{Setup NormalyzerDE statistics object}
\usage{
setupStatisticsObject(designDf, fullDf, sampleCol = "sample",
  conditionCol = "group", batchCol = NULL, logTrans = FALSE,
  leastRepCount = 2)
}
\arguments{
\item{designDf}{Design matrix.}

\item{fullDf}{Full data matrix containing expression data and annotation data.}

\item{sampleCol}{Name of column containing sample IDs.}

\item{conditionCol}{Name of column containing condition levels.}

\item{batchCol}{Optional name of column containing batch effect levels.}

\item{logTrans}{Option for log transforming input data.}

\item{leastRepCount}{Lowest number of replicate required.}
}
\value{
Prepared statistics object.
}
\description{
Setup NormalyzerDE statistics object
}
\examples{
data(example_design)
data(example_stat_data)
setupStatisticsObject(example_design, example_stat_data)
}
