% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerStatistics.R
\name{verifyContrasts}
\alias{verifyContrasts}
\title{Check that a given contrast string is valid given a particular design
matrix. Each level tested for in the contrast should be present in the
condition column for the design matrix.}
\usage{
verifyContrasts(designLevels, contrasts)
}
\arguments{
\item{designLevels}{Vector containing condition levels present in design}

\item{contrasts}{A string containing one or several (comma delimited)
strings for which contrasts should be performed}
}
\value{
None
}
\description{
Mainly meant to verify strings received during server usage.
}
\keyword{internal}
