% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalyzerStatistics.R
\name{verifyContrasts}
\alias{verifyContrasts}
\title{Check that a given contrast string is valid given a particular design
matrix. Each level tested for in the contrast should be present in the
condition column for the design matrix.}
\usage{
verifyContrasts(designLevels, contrastString)
}
\arguments{
\item{designLevels}{Vector containing condition levels present in design}

\item{contrastString}{A string containing one or several (comma delimited)
strings for which contrasts should be performed}
}
\description{
Mainly meant to verify strings received during server usage.
}
\examples{
design <- data.frame(sample=c("s1", "s2", "s3", "s4", "s5", "s6"), 
                     group=c("1", "1", "2", "2", "3", "3"))
contrastString <- "1-2,1-3"
calculateContrasts(design[["group"]], contrastString)
}
