% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{verifySummarizedExperiment}
\alias{verifySummarizedExperiment}
\title{Verify that SummarizedExperiment setup matches the data matrix}
\usage{
verifySummarizedExperiment(summarizedExp, sampleCol)
}
\arguments{
\item{summarizedExp}{SummarizedExperiment object.}

\item{sampleCol}{Column in the SummarizedExperiment colData containing sample IDs.}
}
\value{
None
}
\description{
Verify that SummarizedExperiment setup matches the data matrix
}
\keyword{internal}
