% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{verifySummarizedExperiment}
\alias{verifySummarizedExperiment}
\title{Verify that design matrix setup matches the data matrix}
\usage{
verifySummarizedExperiment(summarizedExp, sampleCol)
}
\arguments{
\item{sampleCol}{Column in design matrix containing sample IDs.}

\item{fullMatrix}{Dataframe with input data.}

\item{designMatrix}{Dataframe with design setup.}
}
\value{
None
}
\description{
Verify that design matrix setup matches the data matrix
}
\keyword{internal}
