% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{verifyValidNumbers}
\alias{verifyValidNumbers}
\title{Verify that input fields conform to the expected formats}
\usage{
verifyValidNumbers(rawDataOnly, groups, quiet = FALSE)
}
\arguments{
\item{rawDataOnly}{Dataframe with input data.}

\item{groups}{Condition levels for comparisons.}
}
\value{
Parsed rawdata where 0 values are replaced with NA
}
\description{
Verify that input fields conform to the expected formats
}
\keyword{internal}
